Fixed user sign in flow and missing sandbox imports Added a navigation container to sandbox screen and moved sandbox screen to a new file
