+++ package.json
@@ -20,8 +20,8 @@
     "@react-navigation/drawer": "^5.12.4",
     "@react-navigation/native": "^5.9.2",
     "@react-navigation/stack": "^5.9.1",
-    "aws-amplify": "^3.4.3",
-    "aws-amplify-react-native": "^4.2.7",
+    "aws-amplify": "^5.0.11",
+    "aws-amplify-react-native": "^7.0.2",
     "aws-sdk": "^2.822.0",
     "expo": "^45.0.0",
     "expo-av": "~11.2.3",
@@ -51,6 +51,7 @@
     "react-native-screens": "~3.11.1",
     "react-native-web": "0.17.7",
     "react-navigation": "^4.4.4",
+    "regenerator-runtime": "^0.13.11",
     "tlds": "^1.221.1"
   },
   "devDependencies": {
